Started working on the IE fixes for UA

+++ src/test/java/com/remedy/resources/DriverScript.java
@@ -124,7 +124,6 @@ public class DriverScript {
 			System.setProperty("webdriver.ie.driver",
 					IEDrvrPath + "IEDriverServer_Win32" + File.separator + "IEDriverServer.exe");
 			driver = new InternetExplorerDriver(caps);
-
 			break;
 		case "phantomJS":
 			String phantomJSDrvrPath;
@@ -180,8 +179,17 @@ public class DriverScript {
 	}
 
 	public void quitDriver() {
-		driver.manage().deleteAllCookies();
-		driver.quit();
-		driver = null;
+		if(Config.getProperty("Browser").equals("ie"))
+		{
+			driver.quit();
+			driver = null;
+		}
+		else
+		{
+			driver.manage().deleteAllCookies();
+			driver.quit();
+			driver = null;
+		}
+		
 	}
 }
